switched to c interface for zoltan
